OPEN-64: Cleanup files
