+++ core-modules/crypto/src/ossl/ec.rs
@@ -6,130 +6,6 @@ use std::os::raw::{c_char, c_int};
 use std::ptr;
 use crypto_openssl_sys::point_conversion_form_t::POINT_CONVERSION_UNCOMPRESSED;
 
-pub struct PKeyCtx(pub *mut EVP_PKEY_CTX);
-
-impl PKeyCtx {
-    /// expose the raw EVP_PKEY_CTX pointer
-    pub fn as_ptr(&self) -> *mut EVP_PKEY_CTX {
-        self.0
-    }
-}
-
-impl Drop for PKeyCtx {
-    fn drop(&mut self) {
-        unsafe { EVP_PKEY_CTX_free(self.0) };
-    }
-}
-
-/// BIO wrapper
-pub struct Bio(*mut BIO);
-
-impl Bio {
-    /// Create a new memory BIO
-    pub fn new_mem() -> OsslResult<Self> {
-        let b = unsafe { BIO_new(BIO_s_mem()) };
-        if b.is_null() {
-            Err(openssl_error("Failed to create BIO"))
-        } else {
-            Ok(Bio(b))
-        }
-    }
-
-    /// Create a BIO from a byte slice
-    pub fn from_slice(data: &[u8]) -> OsslResult<Self> {
-        let b = unsafe { BIO_new_mem_buf(data.as_ptr() as *const _, data.len() as c_int) };
-        if b.is_null() {
-            Err(openssl_error("Failed to create BIO from buffer"))
-        } else {
-            Ok(Bio(b))
-        }
-    }
-
-    /// Read all bytes from BIO
-    pub fn to_vec(&self) -> Vec<u8> {
-        unsafe {
-            let mut ptr_data: *mut c_char = std::ptr::null_mut();
-            let len = BIO_get_mem_data(self.0, &mut ptr_data) as isize;
-            if len <= 0 || ptr_data.is_null() {
-                return Vec::new();
-            }
-            // now cast to *const u8
-            let slice = unsafe { std::slice::from_raw_parts(ptr_data as *const u8, len as usize) };
-            slice.to_vec()
-        }
-    }
-}
-
-impl Drop for Bio {
-    fn drop(&mut self) {
-        unsafe { BIO_free(self.0) };
-    }
-}
-
-/// EVP_PKEY wrapper
-pub struct PKey(pub *mut EVP_PKEY);
-
-impl PKey {
-    pub fn from_der_private(data: &[u8]) -> OsslResult<Self> {
-        let bio = Bio::from_slice(data)?;
-        let p = unsafe { d2i_PrivateKey_bio(bio.0, ptr::null_mut()) };
-        if p.is_null() {
-            Err(openssl_error("Failed to load private key from DER"))
-        } else {
-            Ok(PKey(p))
-        }
-    }
-
-    pub fn from_der_public(data: &[u8]) -> OsslResult<Self> {
-        let bio = Bio::from_slice(data)?;
-        let p = unsafe { d2i_PUBKEY_bio(bio.0, ptr::null_mut()) };
-        if p.is_null() {
-            Err(openssl_error("Failed to load public key from DER"))
-        } else {
-            Ok(PKey(p))
-        }
-    }
-
-    pub fn to_der_private(&self) -> OsslResult<Vec<u8>> {
-        let bio = Bio::new_mem()?;
-        ossl_check!(
-            unsafe {
-                i2d_PKCS8PrivateKey_bio(
-                    bio.0,
-                    self.0,
-                    ptr::null(),
-                    ptr::null(),
-                    0,
-                    None,
-                    ptr::null_mut(),
-                )
-            },
-            "Failed to convert private key to DER"
-        );
-        Ok(bio.to_vec())
-    }
-
-    pub fn to_der_public(&self) -> OsslResult<Vec<u8>> {
-        let bio = Bio::new_mem()?;
-        ossl_check!(
-            unsafe { i2d_PUBKEY_bio(bio.0, self.0) },
-            "Failed to convert public key to DER"
-        );
-        Ok(bio.to_vec())
-    }
-
-    /// expose the raw EVP_PKEY pointer
-    pub fn as_ptr(&self) -> *mut EVP_PKEY {
-        self.0
-    }
-}
-
-impl Drop for PKey {
-    fn drop(&mut self) {
-        unsafe { EVP_PKEY_free(self.0) };
-    }
-}
-
 /// EC point wrapper
 pub struct EcPoint {
     group: *mut EC_GROUP,
