+++ core-modules/crypto/src/ossl/mod.rs
@@ -4,3 +4,5 @@ pub(crate) mod digest;
 pub(crate) mod mac;
 pub(crate) mod mlkem;
 pub(crate) mod cipher;
+mod bio;
+mod key;
